+++ src/querybuilder/components/PromQueryEditorSelector.tsx
@@ -30,7 +30,7 @@ import { promQueryModeller } from '../PromQueryModeller';
 import { buildVisualQueryFromString } from '../parsing';
 import { QueryEditorModeToggle } from '../shared/QueryEditorModeToggle';
 import { QueryHeaderSwitch } from '../shared/QueryHeaderSwitch';
-import { promQueryEditorExplainKey, promQueryEditorRawQueryKey, useFlag } from '../shared/hooks/useFlag';
+import { promQueryEditorExplainKey, useFlag } from '../shared/hooks/useFlag';
 import { QueryEditorMode } from '../shared/types';
 import { changeEditorMode, getQueryWithDefaults } from '../state';
 
@@ -46,7 +46,6 @@ export const PromQueryEditorSelector = React.memo<Props>((props) => {
   const [dataIsStale, setDataIsStale] = useState(false);
   const [trace, setTrace] = useState(false);
   const { flag: explain, setFlag: setExplain } = useFlag(promQueryEditorExplainKey);
-  const { flag: rawQuery, setFlag: setRawQuery } = useFlag(promQueryEditorRawQueryKey, true);
 
   const query = getQueryWithDefaults(props.query, app);
   // This should be filled in from the defaults by now.
@@ -77,11 +76,6 @@ export const PromQueryEditorSelector = React.memo<Props>((props) => {
   useEffect(() => {
     setDataIsStale(false);
   }, [data]);
-
-  const onQueryPreviewChange = (event: SyntheticEvent<HTMLInputElement>) => {
-    const isEnabled = event.currentTarget.checked;
-    setRawQuery(isEnabled);
-  };
 
   const onChangeInternal = (query: PromQuery) => {
     setDataIsStale(true);
@@ -131,9 +125,6 @@ export const PromQueryEditorSelector = React.memo<Props>((props) => {
 
         <QueryHeaderSwitch label="Explain" value={explain} onChange={onShowExplainChange}/>
         <QueryHeaderSwitch label="Trace" value={trace} onChange={onShowTracingChange}/>
-        {editorMode === QueryEditorMode.Builder && (
-            <QueryHeaderSwitch label="Raw query" value={rawQuery} onChange={onQueryPreviewChange}/>
-        )}
         <FlexItem grow={1}/>
         {app !== CoreApp.Explore && (
           <Button
@@ -166,7 +157,7 @@ export const PromQueryEditorSelector = React.memo<Props>((props) => {
             onChange={onChangeInternal}
             onRunQuery={props.onRunQuery}
             data={data}
-            showRawQuery={rawQuery}
+            showRawQuery={true}
             showExplain={explain}
             showTrace={trace}
           />